requirements.txt
appdirs==1.4.4
asgiref==3.3.1
autopep8==1.5.5
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.1
Django==3.0.5
django-filter==2.4.0
djangorestframework==3.12.2
djongo==1.3.4
drf-yasg==1.20.0
filelock==3.0.12
idna==2.10
inflection==0.5.1
install==1.3.4
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.1
ordered-set==4.0.2
packaging==20.9
pandas==1.2.2
pipenv==2020.11.15
pycodestyle==2.6.0
pymongo==3.11.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.2.4
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.3
virtualenv==20.4.2
virtualenv-clone==0.5.4
